+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * feat: Add support for launching dockerized local networks (#233)
 * fix: When deleting a canister, also delete the id from the id store.
 * chore!: rename `icp network run` to `icp network start
+* feat: install Candid UI canister after starting a local network
 
 # v0.1.0-beta.1